448: Updating common.Links definition to match latest schema def
The Links definition in the common package was left as an old representation (which produced the same json as the latest version).
Updating the common def to make future code generation runs easier. Removing the definition from the vrp package and reusing the common def for vrps.

Bumping project version to 1.1.4-SNAPSHOT

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/448