575: Changing HttpStatus for the Initiation & Risk validations
Issues: https://github.com/secureapigateway/secureapigateway/issues/575
Description: Changed the HttpStatus for both Initiation & Risk validations from "FORBIDDEN" to "BAD_REQUEST" according to the spec.
E.g.: "The PISP must ensure that the Initiation and Risk sections of the international-payment match the corresponding Initiation and Risk sections of the international-payment-consent resource. If the two do not match, the ASPSP must not process the request and must respond with a 400 (Bad Request)."
https://openbankinguk.github.io/read-write-api-site3/v3.1.4/resources-and-data-models/pisp/international-payments.html#post-international-payments
These requirements apply to all Payment APIs except for File Payments which does not contain both Initiation & Risk objects, it only contains the Initiation object.