1032: Adding FR data model classes for VRP consents
FR data model classes for VRP consents are required in order to store them in the RCS Consent Store.

Adding FRDomesticVRPConsentConverters to handle conversion between FR and OB types.

Fixing OBDomesticVRPControlParameters generated code to use the OBSupplementaryData1 type for the supplementaryData field. This is for consistency with other payments which reuse this common type.

Adding OBDomesticVRPControlParametersTest.java

https://github.com/SecureApiGateway/SecureApiGateway/issues/1032