1243: Re-generating events data model using latest code-gen plugin version (#154)
Removing Events Callback API types. These types are not in the schema, we do not offer the Callback API externally (there's no IG route) and it is a maintenance burden for something which is not testable.

Configuring the code-gen plugin to use our common data model objects for certain types e.g. Links

https://github.com/SecureApiGateway/SecureApiGateway/issues/1243